tutorials/Learning Series 8/requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.8
scikit-learn>=1.3
umap-learn>=0.5
tqdm>=4.66
pyyaml>=6.0
